requirements.txt
asgiref==3.2.10
Django==3.1.1
django-extensions==3.1.0
djangorestframework==3.12.1
mysqlclient==2.0.1
pytz==2020.4
sqlparse==0.4.1
whitenoise==5.2.0
django-select2==7.6.1
psycopg2==2.8.6
django-debug-toolbar~=3.2